requirements.txt
CoAPthon==4.0.2
pycrypto==2.6.1
